Added an Interface and Builder class for Retrofit server calls as well as a Repository for handling the data.
I then used this to fetch the data and feed it to a scoped ViewModel.

I use the ViewModel in MainActivity to propagate the data back through my Compose functions where I populate their fields with the relevant data.

I also finished setting up the UI with a TextCard Compose function to display the movie detail fields.